fix refreshing tables with pinned ents, saving used conns, and import .sql files
