+++ apps/studio/src/common/appdb/models/used_connection.ts
@@ -38,6 +38,7 @@ export class UsedConnection extends DbConnectionBase implements ISimpleConnectio
       // TEMP (@day): this is just till we fix the used conn duplication issue
       this.authId = other.authId
       this.libsqlOptions = other.libsqlOptions
+
     }
 
     return this;
