+++ apps/studio/src/components/sidebar/core/FavoriteList.vue
@@ -233,7 +233,7 @@ export default {
         this.$root.$emit(AppEvent.promptQueryImport)
     },
     importFromComputer() {
-      this.$root.$emit(AppEvent.promptQueryImportFromComputer)
+      this.$root.$emit(AppEvent.promptSqlFilesImport)
     },
     maybeUnselect(e) {
       if (!this.selected) return
